+++ pcs/src/main/scala/consumers/no_registral/objeto/application/cqrs/commands/ObjetoUpdateFromObligacionTreintaProcientoHandler.scala
@@ -12,16 +12,40 @@ import consumers.no_registral.tranferencia.infrastructure.dependency_injection.O
 import cqrs.untyped.command.CommandHandler.SyncCommandHandler
 import ddd.eventCounterMax
 import design_principles.actor_model.Response
+import design_principles.actor_model.mechanism.DeliveryIdManagement.isIdempotentInternally
 
 import scala.util.{Success, Try}
 
 class ObjetoUpdateFromObligacionTreintaProcientoHandler(actor: ObjetoActor, requeriment: MonitoringAndMessageProducer)
-  extends SyncCommandHandler[ObjetoCommands.ObjetoUpdateFromObnTreintaPorciento] {
+    extends SyncCommandHandler[ObjetoCommands.ObjetoUpdateFromObnTreintaPorciento] {
   override def handle(
-                       command: ObjetoCommands.ObjetoUpdateFromObnTreintaPorciento
-                     ): Try[Response.SuccessProcessing] = {
+      command: ObjetoCommands.ObjetoUpdateFromObnTreintaPorciento
+  ): Try[Response.SuccessProcessing] = {
     val sender = actor.context.sender()
-    val obj_default: ObjetosTri = ObjetosTri(Some("None"), 0, "None", "None", "None", Some("None"), Some("None"), Some("None"), None, None, Some("None"), None, Some(0), Some("None"), Some(0), Some("None"), Some("None"), Some("None"), Some("None"), Some("None"),None,None)
+    val obj_default: ObjetosTri = ObjetosTri(
+      Some("None"),
+      0,
+      "None",
+      "None",
+      "None",
+      Some("None"),
+      Some("None"),
+      Some("None"),
+      None,
+      None,
+      Some("None"),
+      None,
+      Some(0),
+      Some("None"),
+      Some(0),
+      Some("None"),
+      Some("None"),
+      Some("None"),
+      Some("None"),
+      Some("None"),
+      None,
+      None
+    )
 
     log.debug(
       f"""|CUMBIA
@@ -53,20 +77,32 @@ class ObjetoUpdateFromObligacionTreintaProcientoHandler(actor: ObjetoActor, requ
     implicit val ac: ActorSystem = actor.context.system
     val vinculoActor: ActorRef = ObjetoVinculoActor.startWithRequirements(requeriment)
 
-    actor.persistEvent(event) { () =>
-      actor.state += event
-      //if (initialization != "true")
-      //  actor.informParent(command, actor.state)
-      if (actor.state.eventCounter == eventCounterMax) {
-        actor.deleteSnapshots(SnapshotSelectionCriteria(actor.lastSequenceNr - 200))
-        actor.saveSnapshot(actor.state.copy(eventCounter = 0))
+    if (isIdempotentInternally(command, actor.state.lastDeliveryIdByEvents)) {
+      log.error(
+        s"[${actor.name} | ${actor.persistenceId}] -objeto- respond idempotent because of old delivery id | $command -> " + command.deliveryId + " <= " + actor.state.lastDeliveryIdByEvents
+      )
+      sender ! Response.SuccessProcessing(command.aggregateRoot, command.deliveryId)
+
+    } else {
+      actor.persistEvent(event) { () =>
+        actor.state += event
+        //if (initialization != "true")
+        //  actor.informParent(command, actor.state)
+        if (actor.state.eventCounter == eventCounterMax) {
+          actor.deleteSnapshots(SnapshotSelectionCriteria(actor.lastSequenceNr - 200))
+          actor.saveSnapshot(actor.state.copy(eventCounter = 0))
+        }
+        SendObjetoToObjetoVinculo(vinculoActor,
+                                  actor,
+                                  command.sujetoId,
+                                  command.objetoId,
+                                  command.tipoObjeto,
+                                  actor.state.registro.getOrElse(obj_default).SOJ_ESTADO,
+                                  requeriment,
+                                  command)
+
       }
-
-      SendObjetoToObjetoVinculo(vinculoActor, actor, command.sujetoId, command.objetoId, command.tipoObjeto, actor.state.registro.getOrElse(obj_default).SOJ_ESTADO, requeriment, command)
-
     }
-
     Success(Response.SuccessProcessing(command.aggregateRoot, command.deliveryId))
   }
 }
-
