+++ pcs/src/main/scala/consumers/no_registral/sujeto/application/cqrs/commands/SujetoUpdateFromObjetoAntHandler.scala
@@ -7,6 +7,7 @@ import consumers.no_registral.sujeto.infrastructure.dependency_injection.SujetoA
 import cqrs.untyped.command.CommandHandler.SyncCommandHandler
 import ddd.eventCounterMax
 import design_principles.actor_model.Response
+import design_principles.actor_model.mechanism.DeliveryIdManagement.isIdempotentInternally
 
 import scala.util.{Success, Try}
 
@@ -30,14 +31,19 @@ class SujetoUpdateFromObjetoAntHandler(actor: SujetoActor) extends SyncCommandHa
       command.clasificacionObjeto
     )
 
-    actor.persistEvent(event) { () =>
-      actor.state += event
+    if (isIdempotentInternally(command, actor.state.lastDeliveryIdByEvents)) {
+      println(s"[${actor.name} | ${actor.persistenceId}] respond idempotent because of old delivery id | $command")
+      sender ! Response.SuccessProcessing(command.aggregateRoot, command.deliveryId)
+    } else {
+      actor.persistEvent(event) { () =>
+        actor.state += event
 
-      if (actor.state.eventCounter == eventCounterMax) {
-        actor.deleteSnapshots(SnapshotSelectionCriteria(actor.lastSequenceNr - 200))
-        actor.saveSnapshot(actor.state.copy(eventCounter = 0))
-      }
-      actor.persistSnapshot() { _ =>
+        if (actor.state.eventCounter == eventCounterMax) {
+          actor.deleteSnapshots(SnapshotSelectionCriteria(actor.lastSequenceNr - 200))
+          actor.saveSnapshot(actor.state.copy(eventCounter = 0))
+        }
+        actor.persistSnapshot() { _ =>
+        }
       }
     }
     Success(Response.SuccessProcessing(command.aggregateRoot, command.deliveryId))
